+++ CHANGELOG.md
@@ -12,6 +12,8 @@ and this project will adhere to [Semantic Versioning](https://semver.org/spec/v2
 - rotation angle for Rect
 - rotate method for Point
 - Point.from now accepts PointArray
+- Rect.rotate
+- Point.array
 
 ### Changed